--- components/Tab/Tab.js
@@ -1,1 +1,0 @@
-export default function Tab({ children, isActive }) {}
