+++ styles.js
@@ -10,6 +10,9 @@ export default createGlobalStyle`
 :root {
   --navbar-icon-active: #F86F03;
   --navbar-icon-inactive: #D2E9E9;
+
+  --tab-icon-active: #FF9B50;
+  --tab-icon-inactive: black;
 }
